// tsp-output/server/js/src/generated/models/all/syncly-service.ts
export interface ErrorResponse {
  code: string;

  message: string;

  details?: string;
}
// ...
export interface ParticipantAnswer {
  id: string;

  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];

  createdAt: Temporal.Instant;

  updatedAt: Temporal.Instant;
}
// ...
export interface CreateParticipantRequest {
  name: string;

  canJoinRemotely: boolean;

  comment?: string;

  availableCandidateIds: string[];
}

export interface UpdateParticipantRequest {
  name?: string;

  canJoinRemotely?: boolean;

  comment?: string;

  availableCandidateIds?: string[];
}
// ...
export const ParticipantAnswer = {
  toJsonObject(input: ParticipantAnswer): any {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: input.createdAt.toString(),
      updatedAt: input.updatedAt.toString(),
    };
  },
  fromJsonObject(input: any): ParticipantAnswer {
    return {
      id: input.id,
      name: input.name,
      canJoinRemotely: input.canJoinRemotely,
      comment: input.comment,
      availableCandidateIds: input.availableCandidateIds,
      createdAt: Temporal.Instant.from(input.createdAt),
      updatedAt: Temporal.Instant.from(input.updatedAt),
    };
  },
} as const;